scraping/requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.4
idna==2.8
Js2Py==0.66
pyjsparser==2.7.1
pymongo==3.8.0
python-engineio==3.8.2.post1
python-socketio==4.2.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.3
tzlocal==2.0.0
urllib3==1.25.3
websocket-client==0.56.0
